refactor: move AvailabilityChecker to utils.py
Move AvailabilityChecker class from matchmaker.py to utils.py
for better code organization and maintainability.

Changes:
- Move AvailabilityChecker class to utils.py (line 436-665)
- Add List to typing imports in utils.py
- Update imports in matchmaker.py
- Update imports in availability_conflict_resolver.py
- Update imports in availability_conflict_view.py

Benefits:
- Centralized location for all availability utilities
- Better separation of concerns (utils vs. matchmaking logic)
- Easier to reuse across different modules
- All availability functions now in one place:
  * AvailabilityChecker class
  * parse_availability()
  * intersect_availability()
  * validate_time_range()
  * get_default_availability()

No functional changes - purely organizational refactoring.
