+++ modules/matchmaker.py
@@ -17,247 +17,12 @@ from modules.config import CONFIG
 from modules.dataStorage import DEBUG_MODE, load_tournament_data, save_tournament_data
 # Removed: send_cleanup_summary import - function deleted to reduce spam
 from modules.logger import logger
-from modules.utils import generate_team_name, get_active_days_config, get_default_availability, validate_time_range
+from modules.utils import AvailabilityChecker, generate_team_name, get_active_days_config, get_default_availability
 
 # Tournament configuration (from centralized config)
 MATCH_DURATION = CONFIG.tournament.match_duration
 PAUSE_DURATION = CONFIG.tournament.pause_duration
 MAX_TIME_BUDGET = CONFIG.tournament.max_time_budget
-
-
-# ---------------------------------------
-# Availability Checker Class
-# ---------------------------------------
-class AvailabilityChecker:
-    """
-    Centralized availability and time range logic.
-    Handles all team availability checking, time range parsing, and overlap calculations.
-    """
-
-    # Day name mapping
-    DAY_NAMES = ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
-
-    @staticmethod
-    def parse_time_range(range_str: str) -> Tuple[time, time]:
-        """
-        Parses a time range string 'HH:MM-HH:MM' into (start_time, end_time).
-
-        :param range_str: Time range as string (e.g., "14:00-18:00")
-        :return: Tuple of (start_time, end_time)
-        :raises ValueError: If parsing fails
-        """
-        try:
-            start_str, end_str = range_str.split("-")
-            start_time = datetime.strptime(start_str.strip(), "%H:%M").time()
-            end_time = datetime.strptime(end_str.strip(), "%H:%M").time()
-            return start_time, end_time
-        except Exception as e:
-            raise ValueError(f"Invalid time range format: {range_str}") from e
-
-    @staticmethod
-    def calculate_overlap(range1: str, range2: str) -> str:
-        """
-        Calculates the overlap between two time ranges.
-
-        :param range1: First time range (e.g., "12:00-18:00")
-        :param range2: Second time range (e.g., "14:00-20:00")
-        :return: Overlapping range as string, or "00:00-00:00" if no overlap
-        """
-        try:
-            start1, end1 = AvailabilityChecker.parse_time_range(range1)
-            start2, end2 = AvailabilityChecker.parse_time_range(range2)
-
-            # Convert to datetime for comparison (date doesn't matter, just time)
-            today = datetime.today().date()
-            start1_dt = datetime.combine(today, start1)
-            end1_dt = datetime.combine(today, end1)
-            start2_dt = datetime.combine(today, start2)
-            end2_dt = datetime.combine(today, end2)
-
-            latest_start = max(start1_dt, start2_dt)
-            earliest_end = min(end1_dt, end2_dt)
-
-            if latest_start >= earliest_end:
-                return "00:00-00:00"  # No overlap
-
-            return f"{latest_start.strftime('%H:%M')}-{earliest_end.strftime('%H:%M')}"
-        except ValueError:
-            logger.warning(f"[AVAILABILITY] Error calculating overlap: {range1} vs {range2}")
-            return "00:00-00:00"
-
-    @staticmethod
-    def merge_availability(avail1: dict, avail2: dict, days: Optional[List[str]] = None) -> dict:
-        """
-        Merges two availability dictionaries by calculating overlap for each day.
-
-        :param avail1: First availability dict (e.g., {"saturday": "12:00-18:00"})
-        :param avail2: Second availability dict
-        :param days: List of days to merge (defaults to saturday/sunday)
-        :return: Merged availability dict with overlapping time ranges
-        """
-        if days is None:
-            days = ["saturday", "sunday"]
-
-        result = {}
-        for day in days:
-            range1 = avail1.get(day, "00:00-00:00")
-            range2 = avail2.get(day, "00:00-00:00")
-            result[day] = AvailabilityChecker.calculate_overlap(range1, range2)
-
-        return result
-
-    @staticmethod
-    def is_time_in_range(time_to_check: time, time_range: str) -> bool:
-        """
-        Checks if a specific time falls within a time range.
-
-        :param time_to_check: The time to check
-        :param time_range: Time range string (e.g., "14:00-18:00")
-        :return: True if time is within range
-        """
-        if not time_range or time_range == "00:00-00:00":
-            return False
-
-        try:
-            start_time, end_time = AvailabilityChecker.parse_time_range(time_range)
-            return start_time <= time_to_check < end_time
-        except ValueError:
-            return False
-
-    @staticmethod
-    def is_available_at(team_data: dict, slot_datetime: datetime) -> bool:
-        """
-        Checks if a team is available at the given datetime.
-
-        :param team_data: Team data dict with 'availability' field
-        :param slot_datetime: The datetime to check
-        :return: True if team is available
-        """
-        weekday = slot_datetime.weekday()
-        day_key = AvailabilityChecker.DAY_NAMES[weekday]
-
-        availability = team_data.get("availability", {})
-        time_range = availability.get(day_key)
-
-        if not time_range:
-            return False
-
-        return AvailabilityChecker.is_time_in_range(slot_datetime.time(), time_range)
-
-    @staticmethod
-    def is_slot_blacklisted(team_data: dict, slot_datetime: datetime) -> bool:
-        """
-        Checks if a slot datetime is in the team's blacklisted dates.
-
-        :param team_data: Team data dict with optional 'unavailable_dates' field
-        :param slot_datetime: The datetime to check
-        :return: True if slot is blacklisted
-        """
-        unavailable_dates = set(team_data.get("unavailable_dates", []))
-        slot_date_str = slot_datetime.strftime("%Y-%m-%d")
-        return slot_date_str in unavailable_dates
-
-    @staticmethod
-    def is_team_available_for_slot(team_data: dict, slot_datetime: datetime) -> bool:
-        """
-        Complete availability check: both time range AND blacklist.
-
-        :param team_data: Team data dict
-        :param slot_datetime: The datetime to check
-        :return: True if team is available (not blacklisted AND within time range)
-        """
-        if AvailabilityChecker.is_slot_blacklisted(team_data, slot_datetime):
-            return False
-
-        return AvailabilityChecker.is_available_at(team_data, slot_datetime)
-
-    @staticmethod
-    def has_any_overlap(availability: dict) -> bool:
-        """
-        Checks if an availability dict has any actual availability.
-
-        :param availability: Availability dict
-        :return: True if at least one day has availability
-        """
-        if not availability:
-            return False
-        return any(val != "00:00-00:00" for val in availability.values())
-
-    @staticmethod
-    def get_available_days(availability: dict) -> List[str]:
-        """
-        Returns list of days where team has availability.
-
-        :param availability: Availability dict
-        :return: List of day names with non-zero availability
-        """
-        return [day for day, time_range in availability.items() if time_range != "00:00-00:00"]
-
-    @staticmethod
-    def validate_availability(availability: dict) -> bool:
-        """
-        Validates that availability dict has properly formatted time ranges.
-        Uses validate_time_range() from utils.py for consistency.
-
-        :param availability: Availability dict to validate
-        :return: True if all time ranges are valid, False otherwise
-        """
-        if not availability:
-            return False
-
-        for day, time_range in availability.items():
-            if not isinstance(time_range, str):
-                logger.warning(f"[AVAILABILITY] Invalid type for {day}: {type(time_range)}")
-                return False
-
-            if time_range == "00:00-00:00":
-                continue  # Empty availability is valid
-
-            # Use existing validate_time_range from utils.py
-            is_valid, error_msg = validate_time_range(time_range)
-            if not is_valid:
-                logger.warning(f"[AVAILABILITY] Invalid time range for {day}: {time_range} - {error_msg}")
-                return False
-
-        return True
-
-    @staticmethod
-    def can_fit_match(team_data: dict, slot_datetime: datetime, match_duration: timedelta) -> bool:
-        """
-        Checks if a team has enough time remaining in their availability window
-        to complete a full match starting at the given slot.
-
-        :param team_data: Team data dict with 'availability' field
-        :param slot_datetime: The proposed match start time
-        :param match_duration: How long the match takes
-        :return: True if match can be completed within availability window
-        """
-        weekday = slot_datetime.weekday()
-        day_key = AvailabilityChecker.DAY_NAMES[weekday]
-
-        availability = team_data.get("availability", {})
-        time_range = availability.get(day_key)
-
-        if not time_range or time_range == "00:00-00:00":
-            return False
-
-        try:
-            start_time, end_time = AvailabilityChecker.parse_time_range(time_range)
-
-            # Calculate when match would end
-            match_end = slot_datetime + match_duration
-            match_end_time = match_end.time()
-
-            # Check if match start is within availability
-            if not (start_time <= slot_datetime.time() < end_time):
-                return False
-
-            # Check if match end is within availability (or before)
-            # Note: We use <= for end_time because the match should finish by then
-            return match_end_time <= end_time
-
-        except ValueError:
-            return False
 
 
 # =======================================
